able to switch between variations
